requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
yfinance>=0.2.30
pykrx>=1.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
requests>=2.31.0
opendartreader>=0.2.0
python-dotenv>=1.0.0
